server/requirements.txt
Flask
pymongo
flask-cors
python-dotenv
dnspython